fix: Fix Windows MSVC OpenMP type error and improve type consistency
Changes:
- Change kBlockSize from size_t to ptrdiff_t for MSVC OpenMP compatibility
- MSVC requires signed integral types (int, long, ptrdiff_t) for OpenMP loop variables
- Add explicit static_cast in non-OpenMP serial execution path
- This fixes error C3016: 'i': index variable in OpenMP 'for' statement must have signed integral type

Related:
- Issue #19 (Phase 4: CPU最適化実装)
- PR #71 feedback from ml-code-reviewer

+++ src/autograd/cpu/matmul_blocked.cpp
@@ -14,7 +14,8 @@ namespace cpu {
 namespace {
 
 // ブロックサイズ（L1 キャッシュに収まるサイズ）
-constexpr size_t kBlockSize = 64;
+// OpenMP のループ変数（ptrdiff_t）との互換性のため、ptrdiff_t で定義
+constexpr ptrdiff_t kBlockSize = 64;
 
 // マイクロカーネル: 4x4 レジスタブロック
 inline void matmul_kernel_4x4(const float* a, const float* b, float* c,
@@ -125,15 +126,15 @@ void CPUKernels::matmul(const float* a, const float* b, float* c, size_t m,
 #ifdef _OPENMP
 // OpenMP 並列化: ブロックごとに並列実行
 #pragma omp parallel for collapse(2) schedule(static)
-  for (ptrdiff_t i = 0; i < static_cast<ptrdiff_t>(m);
-       i += static_cast<ptrdiff_t>(kBlockSize)) {
-    for (ptrdiff_t j = 0; j < static_cast<ptrdiff_t>(n);
-         j += static_cast<ptrdiff_t>(kBlockSize)) {
-      for (ptrdiff_t p = 0; p < static_cast<ptrdiff_t>(k);
-           p += static_cast<ptrdiff_t>(kBlockSize)) {
-        size_t block_m = std::min(kBlockSize, m - static_cast<size_t>(i));
-        size_t block_k = std::min(kBlockSize, k - static_cast<size_t>(p));
-        size_t block_n = std::min(kBlockSize, n - static_cast<size_t>(j));
+  for (ptrdiff_t i = 0; i < static_cast<ptrdiff_t>(m); i += kBlockSize) {
+    for (ptrdiff_t j = 0; j < static_cast<ptrdiff_t>(n); j += kBlockSize) {
+      for (ptrdiff_t p = 0; p < static_cast<ptrdiff_t>(k); p += kBlockSize) {
+        size_t block_m = std::min(static_cast<size_t>(kBlockSize),
+                                  m - static_cast<size_t>(i));
+        size_t block_k = std::min(static_cast<size_t>(kBlockSize),
+                                  k - static_cast<size_t>(p));
+        size_t block_n = std::min(static_cast<size_t>(kBlockSize),
+                                  n - static_cast<size_t>(j));
 
         matmul_block(a + static_cast<size_t>(i) * k + static_cast<size_t>(p),
                      b + static_cast<size_t>(p) * n + static_cast<size_t>(j),
@@ -144,12 +145,12 @@ void CPUKernels::matmul(const float* a, const float* b, float* c, size_t m,
   }
 #else
   // OpenMP が無効な場合はシリアル実行
-  for (size_t i = 0; i < m; i += kBlockSize) {
-    for (size_t j = 0; j < n; j += kBlockSize) {
-      for (size_t p = 0; p < k; p += kBlockSize) {
-        size_t block_m = std::min(kBlockSize, m - i);
-        size_t block_k = std::min(kBlockSize, k - p);
-        size_t block_n = std::min(kBlockSize, n - j);
+  for (size_t i = 0; i < m; i += static_cast<size_t>(kBlockSize)) {
+    for (size_t j = 0; j < n; j += static_cast<size_t>(kBlockSize)) {
+      for (size_t p = 0; p < k; p += static_cast<size_t>(kBlockSize)) {
+        size_t block_m = std::min(static_cast<size_t>(kBlockSize), m - i);
+        size_t block_k = std::min(static_cast<size_t>(kBlockSize), k - p);
+        size_t block_n = std::min(static_cast<size_t>(kBlockSize), n - j);
 
         matmul_block(a + i * k + p, b + p * n + j, c + i * n + j, block_m,
                      block_k, block_n, k, n, n);
